Address PR review: fix clippy/module-size, split usd module, docs, changeset
CI (rust-tests) fixes:
- Fix clippy needless_range_loop in the USD tests (zip instead of 0..3 index)
  and remove a dead `vcount` (unused_variable) left by the chunks_exact refactor.
- Split usd.rs (586 lines) under the 400-line module-size ratchet into the
  module root (orchestration + mesh gate), usd/emit.rs (prim/mesh/material/
  attribute emission) and usd/fmt.rs (identifier/number/colour helpers + Namer).

Review-comment fixes:
- IFC EXPRESS attribute leaves use exact PascalCase: ifc:GlobalId / ifc:Name /
  ifc:Description (ifc:class stays lower-case — it is the IFC type, not an
  EXPRESS attribute).
- CLI: ifcx/usd are whole-model exports, so a --type/--where/--storey/--limit
  filter is now ignored (with a note) instead of aborting the export on a
  zero-match or invalid filter.
- Add a unit test for the colliding-property-name dedup path in emit_attributes.
- Document the USD surface: --format usd in the CLI guide, export_usd in the MCP
  guide, and a dedicated OpenUSD section + GeometryProcessor.exportUsd row in the
  exporting guide.
- Add a changeset (minor: @ifc-lite/geometry, @ifc-lite/cli, @ifc-lite/mcp).
